requirements.txt
# Core app dependencies
fastapi==0.115.2
uvicorn[standard]==0.34.2

# Database & ORM
asyncpg==0.30.0
sqlalchemy==2.0.40
psycopg2-binary==2.9.10

# Data models & I/O
pydantic==2.11.4
python-dotenv==1.1.0
pyyaml==6.0.2

# Monitoring
prometheus-client==0.21.1

# Math / Data
numpy==2.2.5
pandas==2.2.3
scipy==1.15.2
statsmodels==0.14.4
scikit-learn==1.6.1

# Network & APIs
aiohttp==3.11.18
websockets==15.0.1
requests==2.32.3
